Conflicts:
	SOAP/SOAP/Scripts/Control.js
	SOAP/SOAP/Views/Home/Administration.cshtml

+++ SOAP/SOAP/Models/Callbacks/AnestheticConcernCallback.cs
@@ -12,13 +12,13 @@ namespace SOAP.Models.Callbacks
             AnesthesiaConcern anestheticConcern = new AnesthesiaConcern();
             anestheticConcern.Id = Convert.ToInt32(read["a.Id"]);
             anestheticConcern.PatientId = Convert.ToInt32(read["a.PatientId"].ToString());
-            anestheticConcern.Concern.Id = Convert.ToInt16(read["a.ConcernId"].ToString());
+            anestheticConcern.Concern.Id = Convert.ToInt32(read["a.ConcernId"].ToString());
 
             foreach (AnesthesiaConcern.LazyComponents a in lazyComponents)
             {
                 if (a == AnesthesiaConcern.LazyComponents.LOAD_CONCERN_WITH_DETAILS)
                 {
-                    anestheticConcern.Concern.Category.Id = Convert.ToInt16(read["b.CategoryId"].ToString());
+                    anestheticConcern.Concern.Category.Id = Convert.ToInt32(read["b.CategoryId"].ToString());
                     anestheticConcern.Concern.Label = read["b.Label"].ToString();
                     anestheticConcern.Concern.OtherFlag = Convert.ToChar(read["b.OtherFlag"].ToString());
                     anestheticConcern.Concern.Description = read["b.Description"].ToString();
